Moved dataframe functionality to main module. Added a demo of it to GSQL101.
