+++ pyTigerGraph.egg-info/SOURCES.txt
@@ -2,7 +2,6 @@ README.md
 setup.cfg
 setup.py
 pyTigerGraph/__init__.py
-pyTigerGraph/graphToDataFrame.py
 pyTigerGraph/gsql.py
 pyTigerGraph/pyTigerGraph.py
 pyTigerGraph.egg-info/PKG-INFO
